Add caption manager
